Refactor CRUD into separate modules
