+++ Backend/crud.py
@@ -24,297 +24,9 @@ import schemas # Importa todos os schemas
 # Configuração básica de logging
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
-
-
-# --- User CRUD ---
-def get_user(db: Session, user_id: int) -> Optional[User]:
-    return db.query(User).filter(User.id == user_id).first()
-
-def get_user_by_email(db: Session, email: str) -> Optional[User]:
-    return db.query(User).filter(User.email == email).first()
-
-def get_users(db: Session, skip: int = 0, limit: int = 100) -> List[User]:
-    return db.query(User).offset(skip).limit(limit).all()
-
-def create_user(db: Session, user: schemas.UserCreate) -> User:
-    hashed_password = security.get_password_hash(user.password)
-    
-    # Define os limites iniciais baseados nas configurações padrão para usuários sem plano
-    default_limite_produtos = settings.DEFAULT_LIMIT_PRODUTOS_SEM_PLANO
-    default_limite_enriquecimento = settings.DEFAULT_LIMIT_ENRIQUECIMENTO_SEM_PLANO
-    default_limite_geracao_ia = settings.DEFAULT_LIMIT_GERACAO_IA_SEM_PLANO
-    
-    db_user = User(
-        email=user.email,
-        hashed_password=hashed_password,
-        nome_completo=user.nome_completo,
-        idioma_preferido=user.idioma_preferido,
-        chave_openai_pessoal=user.chave_openai_pessoal,
-        chave_google_gemini_pessoal=user.chave_google_gemini_pessoal,
-        limite_produtos=default_limite_produtos,
-        limite_enriquecimento_web=default_limite_enriquecimento,
-        limite_geracao_ia=default_limite_geracao_ia,
-        is_active=True, # Usuários são ativos por padrão na criação normal
-        is_superuser=False # Superuser só é definido explicitamente
-    )
-
-    if user.plano_id:
-        plano = get_plano(db, plano_id=user.plano_id)
-        if plano:
-            db_user.plano_id = plano.id
-            db_user.limite_produtos = plano.limite_produtos
-            db_user.limite_enriquecimento_web = plano.limite_enriquecimento_web
-            db_user.limite_geracao_ia = plano.limite_geracao_ia
-            # data_expiracao_plano pode ser definida aqui se for relevante (ex: +30 dias)
-            # db_user.data_expiracao_plano = datetime.now(timezone.utc) + timedelta(days=30)
-        else:
-            logger.warning(f"Plano com ID {user.plano_id} não encontrado ao criar usuário {user.email}. Usando defaults.")
-            # Mantém os limites default caso o plano não seja encontrado (embora a validação deva ocorrer no router)
-
-    db.add(db_user)
-    db.commit()
-    db.refresh(db_user)
-    return db_user
-
-def update_user(db: Session, db_user: User, user_update: Union[schemas.UserUpdate, schemas.UserUpdateByAdmin, schemas.UserUpdateOAuth]) -> User:
-    update_data = user_update.model_dump(exclude_unset=True)
-
-    if "password" in update_data and update_data["password"]:
-        hashed_password = security.get_password_hash(update_data["password"])
-        db_user.hashed_password = hashed_password
-        del update_data["password"] # Evita que seja atribuído diretamente no loop
-
-    # Atualizar plano e limites associados
-    if "plano_id" in update_data:
-        new_plano_id = update_data["plano_id"]
-        if new_plano_id is None: # Usuário está removendo o plano
-            db_user.plano_id = None
-            db_user.limite_produtos = settings.DEFAULT_LIMIT_PRODUTOS_SEM_PLANO
-            db_user.limite_enriquecimento_web = settings.DEFAULT_LIMIT_ENRIQUECIMENTO_SEM_PLANO
-            db_user.limite_geracao_ia = settings.DEFAULT_LIMIT_GERACAO_IA_SEM_PLANO
-            db_user.data_expiracao_plano = None
-        else:
-            plano = get_plano(db, plano_id=new_plano_id)
-            if plano:
-                db_user.plano_id = plano.id
-                db_user.limite_produtos = plano.limite_produtos
-                db_user.limite_enriquecimento_web = plano.limite_enriquecimento_web
-                db_user.limite_geracao_ia = plano.limite_geracao_ia
-                # db_user.data_expiracao_plano = datetime.now(timezone.utc) + timedelta(days=30) # ou outra lógica
-            else:
-                # Se o plano não for encontrado, talvez não atualizar ou logar um erro.
-                # Por ora, não altera o plano se o ID for inválido e não for None.
-                logger.warning(f"Plano com ID {new_plano_id} não encontrado ao atualizar usuário {db_user.email}. Plano não alterado.")
-        # Remove plano_id de update_data para não ser processado no loop abaixo,
-        # pois os limites são tratados separadamente aqui ou via UserUpdateByAdmin.
-        if "plano_id" in update_data: # precisa verificar novamente pois UserUpdateOAuth pode não ter
-            del update_data["plano_id"]
-
-
-    # Para campos que podem ser atualizados diretamente
-    for field, value in update_data.items():
-        if hasattr(db_user, field):
-            setattr(db_user, field, value)
-
-    db.commit()
-    db.refresh(db_user)
-    return db_user
-
-
-def delete_user(db: Session, db_user: User) -> User:
-    db.delete(db_user)
-    db.commit()
-    return db_user
-
-def create_user_oauth(db: Session, user_oauth: schemas.UserCreateOAuth, plano_id_default: Optional[int] = None) -> User:
-    db_user = User(
-        email=user_oauth.email,
-        nome_completo=user_oauth.nome_completo,
-        provider=user_oauth.provider,
-        provider_user_id=user_oauth.provider_user_id,
-        is_active=True,
-        idioma_preferido=user_oauth.idioma_preferido, # Adicionado
-        # Limites padrão, serão atualizados se plano_id_default for fornecido
-        limite_produtos=settings.DEFAULT_LIMIT_PRODUTOS_SEM_PLANO,
-        limite_enriquecimento_web=settings.DEFAULT_LIMIT_ENRIQUECIMENTO_SEM_PLANO,
-        limite_geracao_ia=settings.DEFAULT_LIMIT_GERACAO_IA_SEM_PLANO,
-    )
-    if plano_id_default:
-        plano = get_plano(db, plano_id=plano_id_default)
-        if plano:
-            db_user.plano_id = plano.id
-            db_user.limite_produtos = plano.limite_produtos
-            db_user.limite_enriquecimento_web = plano.limite_enriquecimento_web
-            db_user.limite_geracao_ia = plano.limite_geracao_ia
-            # db_user.data_expiracao_plano = datetime.now(timezone.utc) + timedelta(days=30) # Exemplo
-    
-    # Se o Role padrão "user" existir, atribui-lo
-    default_role = get_role_by_name(db, "user")
-    if default_role:
-        db_user.role_id = default_role.id
-
-    db.add(db_user)
-    db.commit()
-    db.refresh(db_user)
-    return db_user
-
-def get_user_by_provider(db: Session, provider: str, provider_user_id: str) -> Optional[User]:
-    return db.query(User).filter(User.provider == provider, User.provider_user_id == provider_user_id).first()
-
-
-def set_user_password_reset_token(db: Session, user: User, token_hash: str, expires_at: datetime) -> None:
-    """Persist the hashed password reset token for a user."""
-    user.reset_password_token = token_hash
-    user.reset_password_token_expires_at = expires_at
-    db.commit()
-    db.refresh(user)
-
-def get_user_by_reset_token(db: Session, token_hash: str) -> Optional[User]:
-    """Retrieve a user by the hashed password reset token."""
-    return db.query(User).filter(User.reset_password_token == token_hash).first()
-
-# --- Role CRUD ---
-def get_role(db: Session, role_id: int) -> Optional[Role]:
-    return db.query(Role).filter(Role.id == role_id).first()
-
-def get_role_by_name(db: Session, name: str) -> Optional[Role]:
-    return db.query(Role).filter(Role.name == name).first()
-
-def get_roles(db: Session, skip: int = 0, limit: int = 10) -> List[Role]:
-    return db.query(Role).offset(skip).limit(limit).all()
-
-def create_role(db: Session, role: schemas.RoleCreate) -> Role:
-    db_role = Role(name=role.name, description=role.description)
-    db.add(db_role)
-    db.commit()
-    db.refresh(db_role)
-    return db_role
-
-# --- Plano CRUD ---
-def get_plano(db: Session, plano_id: int) -> Optional[Plano]:
-    return db.query(Plano).filter(Plano.id == plano_id).first()
-
-def get_plano_by_name(db: Session, nome: str) -> Optional[Plano]:
-    return db.query(Plano).filter(Plano.nome == nome).first()
-
-def get_planos(db: Session, skip: int = 0, limit: int = 10) -> List[Plano]:
-    return db.query(Plano).offset(skip).limit(limit).all()
-
-def create_plano(db: Session, plano: schemas.PlanoCreate) -> Plano:
-    db_plano = Plano(**plano.model_dump())
-    db.add(db_plano)
-    db.commit()
-    db.refresh(db_plano)
-    return db_plano
-
-def update_plano(db: Session, db_plano: Plano, plano_update: schemas.PlanoUpdate) -> Plano:
-    update_data = plano_update.model_dump(exclude_unset=True)
-    for key, value in update_data.items():
-        setattr(db_plano, key, value)
-    db.commit()
-    db.refresh(db_plano)
-    return db_plano
-
-def delete_plano(db: Session, db_plano: Plano):
-    db.delete(db_plano)
-    db.commit()
-    return db_plano
-
-
-# --- Fornecedor CRUD ---
-def create_fornecedor(db: Session, fornecedor: schemas.FornecedorCreate, user_id: int) -> Fornecedor:
-    # Converter objetos HttpUrl para strings antes de passar para o modelo SQLAlchemy
-    fornecedor_data = fornecedor.model_dump()
-    if fornecedor_data.get("site_url"):
-        fornecedor_data["site_url"] = str(fornecedor_data["site_url"]) # Converte HttpUrl para string
-    if fornecedor_data.get("link_busca_padrao"):
-        fornecedor_data["link_busca_padrao"] = str(fornecedor_data["link_busca_padrao"]) # Converte HttpUrl para string
-
-    db_fornecedor = Fornecedor(**fornecedor_data, user_id=user_id) # Passa o dicionário modificado
-    db.add(db_fornecedor)
-    db.commit()
-    db.refresh(db_fornecedor)
-    return db_fornecedor
-
-def get_fornecedor(db: Session, fornecedor_id: int) -> Optional[Fornecedor]:
-    return db.query(Fornecedor).filter(Fornecedor.id == fornecedor_id).first()
-
-def get_fornecedores_by_user(
-    db: Session,
-    user_id: Optional[int] = None, # Tornar opcional para admins
-    is_admin: bool = False,
-    skip: int = 0,
-    limit: int = 10,
-    search: Optional[str] = None
-) -> List[Fornecedor]:
-    query = db.query(Fornecedor)
-    if not is_admin and user_id:
-        query = query.filter(Fornecedor.user_id == user_id)
-    
-    if search:
-        search_term = f"%{search.lower()}%"
-        query = query.filter(
-            or_(
-                func.lower(Fornecedor.nome).ilike(search_term),
-                func.lower(Fornecedor.email_contato).ilike(search_term), # Ajustado para email_contato
-                func.lower(Fornecedor.contato_principal).ilike(search_term) # Ajustado para contato_principal
-            )
-        )
-    return query.order_by(Fornecedor.nome).offset(skip).limit(limit).all()
-
-
-def count_fornecedores_by_user(
-    db: Session,
-    user_id: Optional[int] = None, # Tornar opcional para admins
-    is_admin: bool = False,
-    search: Optional[str] = None
-) -> int:
-    query = db.query(func.count(Fornecedor.id))
-    if not is_admin and user_id:
-        query = query.filter(Fornecedor.user_id == user_id)
-    
-    if search:
-        search_term = f"%{search.lower()}%"
-        query = query.filter(
-            or_(
-                func.lower(Fornecedor.nome).ilike(search_term),
-                func.lower(Fornecedor.email_contato).ilike(search_term),
-                func.lower(Fornecedor.contato_principal).ilike(search_term)
-            )
-        )
-    return query.scalar() or 0
-
-
-# Localize também a função update_fornecedor e aplique lógica semelhante:
-def update_fornecedor(db: Session, db_fornecedor: Fornecedor, fornecedor_update: schemas.FornecedorUpdate) -> Fornecedor:
-    update_data = fornecedor_update.model_dump(exclude_unset=True)
-
-    # Converter objetos HttpUrl para strings antes de aplicar a atualização
-    if "site_url" in update_data and update_data["site_url"] is not None:
-        update_data["site_url"] = str(update_data["site_url"])
-    if "link_busca_padrao" in update_data and update_data["link_busca_padrao"] is not None:
-        update_data["link_busca_padrao"] = str(update_data["link_busca_padrao"])
-
-    for key, value in update_data.items():
-        setattr(db_fornecedor, key, value)
-    db.commit()
-    db.refresh(db_fornecedor)
-    return db_fornecedor
-
-def delete_fornecedor(db: Session, db_fornecedor: Fornecedor) -> Fornecedor:
-    db.delete(db_fornecedor)
-    db.commit()
-    # Se Fornecedor não tiver cascade para produtos, a referência em produto precisa ser tratada (setar para None)
-    # No seu model.py, Produto tem um relacionamento com Fornecedor, mas Fornecedor não tem cascade="all, delete-orphan" para Produto.
-    # Isso significa que deletar um fornecedor não deleta os produtos.
-    # A ForeignKey `fornecedor_id` em `Produto` deve ser `nullable=True` e `ondelete="SET NULL"` no DB,
-    # ou você precisa fazer isso manualmente:
-    # db.query(Produto).filter(Produto.fornecedor_id == db_fornecedor.id).update({"fornecedor_id": None})
-    # db.commit()
-    # No entanto, a lógica atual do seu `models.Produto` para `fornecedor_id` é `nullable=True`
-    # o que é bom. O `ondelete` no DB é o mais robusto.
-    return db_fornecedor
-
+from .crud_users import *
+from .crud_fornecedores import *
+from .crud_produtos import *
 
 # --- ProductType CRUD ---
 def create_product_type(db: Session, product_type_create: schemas.ProductTypeCreate, user_id: Optional[int] = None) -> ProductType:
@@ -567,222 +279,6 @@ def reorder_attribute_template(db: Session, attribute_id: int, direction: str) -
     db.commit()
     db.refresh(attr_to_move)
     return attr_to_move
-
-# --- Produto CRUD ---
-def create_produto(db: Session, produto: schemas.ProdutoCreate, user_id: int) -> Produto:
-    produto_data = produto.model_dump(exclude_unset=True)
-    
-    # Assegura que campos JSON estejam como dicts
-    if 'dynamic_attributes' in produto_data and isinstance(produto_data['dynamic_attributes'], str):
-        try:
-            produto_data['dynamic_attributes'] = json.loads(produto_data['dynamic_attributes'])
-        except json.JSONDecodeError:
-            raise ValueError("dynamic_attributes não é um JSON string válido.")
-
-    if 'dados_brutos_web' in produto_data and isinstance(produto_data['dados_brutos_web'], str):
-        try:
-            produto_data['dados_brutos_web'] = json.loads(produto_data['dados_brutos_web'])
-        except json.JSONDecodeError:
-            raise ValueError("dados_brutos_web não é um JSON string válido.")
-
-    if 'log_enriquecimento_web' in produto_data and isinstance(produto_data['log_enriquecimento_web'], str):
-        try:
-            produto_data['log_enriquecimento_web'] = json.loads(produto_data['log_enriquecimento_web'])
-        except json.JSONDecodeError:
-            raise ValueError("log_enriquecimento_web não é um JSON string válido.")
-
-    db_produto = Produto(**produto_data, user_id=user_id)
-    db.add(db_produto)
-    db.commit()
-    db.refresh(db_produto)
-    return db_produto
-
-
-def get_produto(db: Session, produto_id: int) -> Optional[Produto]:
-    # Usar selectinload para carregar relacionamentos de forma eficiente se sempre forem acessados
-    return db.query(Produto).options(
-        selectinload(Produto.fornecedor),
-        selectinload(Produto.product_type).selectinload(ProductType.attribute_templates)
-    ).filter(Produto.id == produto_id).first()
-
-
-def get_produtos_by_user(
-    db: Session,
-    user_id: Optional[int], # Se None e is_admin=True, busca todos. Se user_id e is_admin=False, busca do usuário.
-    is_admin: bool,
-    skip: int = 0,
-    limit: int = 10,
-    sort_by: Optional[str] = None,
-    sort_order: Optional[str] = "asc",
-    search: Optional[str] = None,
-    fornecedor_id: Optional[int] = None,
-    product_type_id: Optional[int] = None,
-    categoria: Optional[str] = None, # Adicionado
-    status_enriquecimento_web: Optional[StatusEnriquecimentoEnum] = None, # Adicionado
-    status_titulo_ia: Optional[StatusGeracaoIAEnum] = None, # Adicionado
-    status_descricao_ia: Optional[StatusGeracaoIAEnum] = None # Adicionado
-) -> List[Produto]:
-    query = db.query(Produto).options(
-        selectinload(Produto.fornecedor),
-        selectinload(Produto.product_type) # Carrega product_type, mas não seus atributos aqui para a lista
-    )
-
-    if not is_admin:
-        if user_id is None: # Não deveria acontecer se não for admin e não tiver user_id
-            return []
-        query = query.filter(Produto.user_id == user_id)
-
-    if search:
-        search_term = f"%{search.lower()}%"
-        query = query.filter(
-            or_(
-                func.lower(Produto.nome_base).ilike(search_term),
-                func.lower(Produto.nome_chat_api).ilike(search_term),
-                func.lower(Produto.descricao_original).ilike(search_term),
-                func.lower(Produto.descricao_chat_api).ilike(search_term),
-                func.lower(Produto.sku).ilike(search_term),
-                func.lower(Produto.ean).ilike(search_term),
-                func.lower(Produto.marca).ilike(search_term),
-                func.lower(Produto.modelo).ilike(search_term)
-            )
-        )
-    
-    if fornecedor_id is not None:
-        query = query.filter(Produto.fornecedor_id == fornecedor_id)
-    if product_type_id is not None:
-        query = query.filter(Produto.product_type_id == product_type_id)
-    if categoria:
-        query = query.filter(func.lower(Produto.categoria_original).ilike(f"%{categoria.lower()}%")) # Ou categoria_mapeada
-    if status_enriquecimento_web:
-        query = query.filter(Produto.status_enriquecimento_web == status_enriquecimento_web)
-    if status_titulo_ia:
-        query = query.filter(Produto.status_titulo_ia == status_titulo_ia)
-    if status_descricao_ia:
-        query = query.filter(Produto.status_descricao_ia == status_descricao_ia)
-
-    if sort_by:
-        column_to_sort = getattr(Produto, sort_by, None)
-        if column_to_sort:
-            if sort_order.lower() == "desc":
-                query = query.order_by(desc(column_to_sort))
-            else:
-                query = query.order_by(asc(column_to_sort))
-        else: # Fallback ou log de erro se sort_by for inválido
-            query = query.order_by(Produto.id) 
-    else:
-        query = query.order_by(Produto.id) # Ordenação padrão
-
-    return query.offset(skip).limit(limit).all()
-
-
-def count_produtos_by_user(
-    db: Session,
-    user_id: Optional[int],
-    is_admin: bool,
-    search: Optional[str] = None,
-    fornecedor_id: Optional[int] = None,
-    product_type_id: Optional[int] = None,
-    categoria: Optional[str] = None,
-    status_enriquecimento_web: Optional[StatusEnriquecimentoEnum] = None,
-    status_titulo_ia: Optional[StatusGeracaoIAEnum] = None,
-    status_descricao_ia: Optional[StatusGeracaoIAEnum] = None
-) -> int:
-    query = db.query(func.count(Produto.id))
-
-    if not is_admin:
-        if user_id is None:
-            return 0
-        query = query.filter(Produto.user_id == user_id)
-
-    if search:
-        search_term = f"%{search.lower()}%"
-        query = query.filter(
-            or_(
-                func.lower(Produto.nome_base).ilike(search_term),
-                func.lower(Produto.nome_chat_api).ilike(search_term),
-                func.lower(Produto.descricao_original).ilike(search_term),
-                func.lower(Produto.descricao_chat_api).ilike(search_term),
-                func.lower(Produto.sku).ilike(search_term),
-                func.lower(Produto.ean).ilike(search_term),
-                func.lower(Produto.marca).ilike(search_term),
-                func.lower(Produto.modelo).ilike(search_term)
-            )
-        )
-    if fornecedor_id is not None:
-        query = query.filter(Produto.fornecedor_id == fornecedor_id)
-    if product_type_id is not None:
-        query = query.filter(Produto.product_type_id == product_type_id)
-    if categoria:
-        query = query.filter(func.lower(Produto.categoria_original).ilike(f"%{categoria.lower()}%"))
-    if status_enriquecimento_web:
-        query = query.filter(Produto.status_enriquecimento_web == status_enriquecimento_web)
-    if status_titulo_ia:
-        query = query.filter(Produto.status_titulo_ia == status_titulo_ia)
-    if status_descricao_ia:
-        query = query.filter(Produto.status_descricao_ia == status_descricao_ia)
-        
-    count = query.scalar()
-    return count if count is not None else 0
-
-def update_produto(db: Session, db_produto: Produto, produto_update: schemas.ProdutoUpdate) -> Produto:
-    update_data = produto_update.model_dump(exclude_unset=True)
-
-
-    # Lógica para campos JSON (dynamic_attributes, dados_brutos_web, log_enriquecimento_web, imagens_secundarias_urls)
-    # Pydantic deve entregar dict/list se o schema estiver correto (não Json[Type])
-    # O modelo SQLAlchemy aceita dict/list para colunas JSON.
-    for field in ['dynamic_attributes', 'dados_brutos_web', 'log_enriquecimento_web', 'imagens_secundarias_urls']:
-        if field in update_data and update_data[field] is not None:
-            # Se o schema já garante o tipo correto (dict/list), apenas atribua
-            # Se o schema ainda for Json[Type] e vier uma string, precisa de json.loads aqui
-            # Assumindo que o schema.ProdutoUpdate terá Dict/List para esses campos:
-            pass # setattr abaixo cuidará disso
-          
-    for key, value in update_data.items():
-        setattr(db_produto, key, value)
-    
-    db.commit()
-    db.refresh(db_produto)
-    # Recarregar relacionamentos se forem modificados ou para garantir consistência na resposta
-    db.refresh(db_produto, attribute_names=['fornecedor', 'product_type'])
-    if db_produto.product_type: # Para garantir que os atributos do tipo também sejam carregados se product_type for acessado
-        db.refresh(db_produto.product_type, attribute_names=['attribute_templates'])
-    return db_produto
-
-
-def delete_produto(db: Session, db_produto: Produto) -> Produto:
-    # Antes de deletar, pode ser necessário limpar referências em RegistroUsoIA se não houver cascade
-    # No seu modelo, RegistroUsoIA tem cascade="all, delete-orphan" para produto, então está OK.
-    db.delete(db_produto)
-    db.commit()
-    return db_produto
-
-
-async def save_produto_image(db: Session, produto_id: int, file: UploadFile) -> str:
-    """Salva a imagem enviada para um produto e retorna o caminho relativo."""
-    if not file.filename:
-        raise ValueError("Nome do arquivo não fornecido.")
-
-    upload_dir = Path(settings.UPLOAD_DIRECTORY)
-    if not upload_dir.is_absolute():
-        upload_dir = Path(__file__).resolve().parent / upload_dir
-    upload_dir.mkdir(parents=True, exist_ok=True)
-
-    file_extension = Path(file.filename).suffix
-    unique_filename = f"{uuid.uuid4().hex}{file_extension}"
-    file_path = upload_dir / unique_filename
-
-    try:
-        content = await file.read()
-        with open(file_path, "wb") as f_out:
-            f_out.write(content)
-    except Exception as e:
-        raise IOError(f"Não foi possível salvar o arquivo: {e}")
-    finally:
-        await file.close()
-
-    relative = Path(settings.UPLOAD_DIRECTORY) / unique_filename
-    return f"/{relative.as_posix()}"
 
 # --- RegistroUsoIA CRUD ---
 def create_registro_uso_ia(db: Session, registro_uso: schemas.RegistroUsoIACreate) -> RegistroUsoIA:
